requirements.txt
pandas==2.2.2
tensorflow==2.14.0
keras==2.14.0
keras-bert==0.89.0
numpy==1.26.4
scikit-learn==1.4.2
opencc-python-reimplemented==0.1.7
